+++ tests/test_ipc.py
@@ -114,6 +114,7 @@ def _generate_fake_job() -> job.RunningJobInfo:
         url="fake_url",
         token="fake_token",
         accept_arguments=job.JobAcceptArguments(name="", identity="", metadata=""),
+        worker_id="fake_id",
     )
 
 
@@ -353,9 +354,9 @@ async def test_shutdown_no_job():
 
     assert proc.exitcode == 0
     assert not proc.killed
-    assert (
-        start_args.shutdown_counter.value == 0
-    ), "shutdown_cb isn't called when there is no job"
+    assert start_args.shutdown_counter.value == 0, (
+        "shutdown_cb isn't called when there is no job"
+    )
 
 
 async def test_job_slow_shutdown():